+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 
 ## 3.7.40 - 2022-05-03
 
-> {warning} If you are currently running Craft 3.7.39, you will need to perform this update with Composer.
+> {warning} If you are currently running Craft 3.7.39, you will need to [perform this update with Composer](https://craftcms.com/knowledge-base/updating-from-craft-3-7-39).
 
 ### Added
 - Added the “Craft 4 Upgrade” utility.