+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 4.0.2 (WIP)
 
+### Fixed
+- Fixed a PHP deprecation warning that occurred if the `tablePrefix` database connection setting was `null`.
+
 ## 4.0.1 - 2022-05-06
 
 ### Fixed