+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 - Fixed a bug where soft-deleted drafts’ search keywords weren’t getting re-indexed if they were restored.
+- Fixed an error that occurred when transforming an SVG image without specifying a width or height. ([#11122](https://github.com/craftcms/cms/issues/11122))
 
 ## 3.7.41 - 2022-05-06