+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft CMS 4
 
+## 4.0.2 (WIP)
+
+### Fixed
+- Fixed a PHP deprecation warning that occurred if the `tablePrefix` database connection setting was `null`.
+
 ## 4.0.1 - 2022-05-06
 
 ### Changed